chore: improve Breadcrumb exports
